fix(lsp): verify that nodejs runs (again)
Problem:
The logic added in a26437735b08 wasn't added in the other "lsp startup"
module.

Solution:
Add it to the new module also.
